pyproject.toml
[tool.poetry]
name = "barda"
version = "0.1.0"
description = ""
authors = ["Brian Pepple <bpepple@metron.cloud>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyxdg = "^0.28"
questionary = "^1.10.0"
simyan = "^0.13.0"
titlecase = "^2.4"
bs4 = "^0.0.1"
requests = "^2.28.2"
ratelimit = "^2.2.1"
pillow = "^9.4.0"
mysql-connector-python = "^8.0.32"
mokkari = "^2.3.3"
esak = "^1.3.2"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
pre-commit = "^2.20.0"
flake8 = "^6.0.0"
flake8-builtins = "^2.1.0"
flake8-blind-except = "^0.2.1"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
barda = "barda.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = ["PIL", "bs4", "dateutil", "esak", "mokkari", "mysql", "prompt_toolkit", "pytest", "questionary", "ratelimit", "requests", "simyan", "titlecase", "urllib3", "xdg"]

[tool.coverage.run]
source = ["barda"]

[tool.coverage.report]
show_missing = true